refactor: add utils for validation helpers
