+++ client/src/utils/index.ts
@@ -1,0 +1,3 @@
+import * as ValidationUtils from './validation';
+
+export { ValidationUtils };
